fix: Create a hook for managing audio to isolate complex functions from components
